Move system actions out of nav, update paths
