--- src/components/Nav/SystemActionsTiles/RecentFlows.vue
@@ -1,13 +1,0 @@
-<script>
-export default {}
-</script>
-
-<template>
-  <button
-    class="rounded-lg system-action-container system-action-container-large wide d-flex flex-column align-center justify-center"
-  >
-    RECENT FLOWS
-  </button>
-</template>
-
-<style></style>
