--- src/components/Nav/SystemActionsTiles/TenantInfo.vue
@@ -1,13 +1,0 @@
-<script>
-export default {}
-</script>
-
-<template>
-  <button
-    class="rounded-lg system-action-container system-action-container-large d-flex flex-column align-center justify-center"
-  >
-    TENANT
-  </button>
-</template>
-
-<style></style>
